untested refactor
